Updating TaxEngine [1] replaced BigDecimal with float in a majority of classes that previously stored a "BigDecimal"; now only the TaxEngine and ShoppingCart will utilize a BigDecimal when it is performing the monetary calculations needed for taxes and providing totals [2] pushed the rounding of the totalTax on an item (state + import duty) to its own method.  Reworked the logic to ensure that the last decimal digit of 1 or 2 would not cause a negative tax value [3] removed any imports that were not being utilized in testing [4] fixed an issue with the ShoppingCart where the decimal value was not being rounded correctly [5] added some javadoc descriptions

+++ src/ct/products/Product.java
@@ -1,15 +1,13 @@
 package ct.products;
 
-import java.math.BigDecimal;
-
 public class Product
 {
     private final String id, name;
     private final ProductType type;
     private final boolean isImport;
-    private final BigDecimal msrp;
+    private final float msrp;
 
-    public Product(String id, String name, ProductType type, boolean isImport, BigDecimal msrp)
+    public Product(String id, String name, ProductType type, boolean isImport, float msrp)
     {
         this.id = id;
         this.name = name;
@@ -18,7 +16,7 @@ public class Product
         this.msrp = msrp;
     }
 
-    public BigDecimal getMsrp() {
+    public float getMsrp() {
         return msrp;
     }
